pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fprime-ci"
dynamic = ["version"]
description = "F Prime CI Tooling suite"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE.txt"}
keywords = ["fprime", "embedded", "nasa", "flight", "software"]
authors = [
  {name = "Michael Starch", email = "Michael.D.Starch@jpl.nasa.gov"},
  {name = "Thomas Boyer-Chammard", email = "Thomas.Boyer.Chammard@jpl.nasa.gov"},
  {name = "Kevin F. Ortega", email = "Kevin.F.Ortega@jpl.nasa.gov"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "pyserial>=3.5",
    "pluggy>=1.5.0",
    "requests>=2.32.3",
    "pyyaml"
]

[project.urls]
Homepage = "https://fprime.jpl.nasa.gov"
Documentation = "https://fprime.jpl.nasa.gov/docs"
Repository = "https://github.com/fprime-community/fprime-ci"

####
# Entry Points:
#
# Defines the list of entry-level (scripts) that are defined by this package. This allows
# standard use of utilities that ship as part of F prime.
####
[project.scripts]
fprime-ci = "fprime_ci.__main__:main"
fprime-ci-power = "fprime_ci.power:main"

####
# setuptools_scm dynamically generates version number from git, as well as automatically
# include all data files tracked by git (e.g. flash/static/** files).
# See https://setuptools.pypa.io/en/latest/userguide/datafiles.html
####
[tool.setuptools_scm]
